requirements.txt
discord.py[voice]
pymongo[srv]
pillow
asyncpg
apscheduler
giphy_client
numpy
